Move the class DeviceFactoryRequest to a new file
Create a new file for the device factory data structures.
