+++ pyanaconda/modules/storage/partitioning/interactive/change_device.py
@@ -21,7 +21,7 @@ from blivet.size import Size
 from dasbus.structure import compare_data
 
 from pyanaconda.anaconda_loggers import get_module_logger
-from pyanaconda.modules.common.structures.partitioning import DeviceFactoryRequest
+from pyanaconda.modules.common.structures.device_factory import DeviceFactoryRequest
 from pyanaconda.modules.common.task import Task
 from pyanaconda.modules.storage.partitioning.interactive.utils import destroy_device, \
     get_device_factory_arguments, revert_reformat, resize_device, reformat_device, \
